Refactor Footprint traits and pinmap attachment
- Moved pinmap-related traits into the Footprint class
- Removed standalone trait files for equal pins and pinmap attachment
- Updated references in various footprint classes to use new Footprint trait namespacing
- Simplified trait implementations within the Footprint class
