+++ src/faebryk/library/SMDTwoPin.py
@@ -47,5 +47,5 @@ class SMDTwoPin(F.Footprint):
             )
 
     kicad_footprint: _has_kicad_footprint
-    equal_pins: F.has_equal_pins_in_ifs
-    attach_via_pinmap: F.can_attach_via_pinmap_equal
+    equal_pins: F.Footprint.has_equal_pins
+    attach_via_pinmap: F.Footprint.can_attach_via_pinmap_equal
