+++ src/faebryk/library/SOIC.py
@@ -36,5 +36,5 @@ class SOIC(F.Footprint):
             )
 
     kicad_footprint: _has_kicad_footprint
-    attach_via_pinmap: F.can_attach_via_pinmap_equal
-    equal_pins: F.has_equal_pins_in_ifs
+    attach_via_pinmap: F.Footprint.can_attach_via_pinmap_equal
+    equal_pins: F.Footprint.has_equal_pins
